# Optimize MultiHeadAttention forward method