@stephenroller 2021-03-26
heads up @EricMichaelSmith might help with latency on [internal thing]