@stephenroller 2021-03-26
On a volta with bs 1.
With a taskmaster2 BART model:
master: 50 generations in 31.4s
this patch: 50 generations in 29.0s, or about 7% speedup

With blender3b:
master: 50 generations in 74s
this patch: 50 generations in 67s, or about a 9% speedup.

on cpu with bs1 with taskmaster 2 BART model:
master: 50 generations in 59.0s
this patch: 50 generations in 54.2, or about 8% speedup

Real nice find!