@stephenroller 2021-03-26
benchmark script i used:
```python
#!/usr/bin/env python3
import tqdm

from parlai.core.agents import create_agent_from_model_file

agent = create_agent_from_model_file('zoo:blender/blender_3B/model')

import timeit

start = timeit.default_timer()
for i in tqdm.tqdm(range(50)):
    agent.observe(
        {'text': 'Can you find me a mexican restaurant for 2?', 'episode_done': True}
    )
    print(agent.act()['text'])
    agent.reset()
end = timeit.default_timer()

print(end - start)
```