@dejk-alg 2021-03-26
@stephenroller Stephen, thanks for having a look at this! Glad that I could participate here in some way. The logic behind this fix is pretty simple. Let me reiterate on what I said in the PR description a bit. The reason behind incremental state usage is obviously saving keys and values for attention to reuse them. When running inference of any encoder-decoder transformer, on every decoding step we need to get both encoder and self-attention keys and values for each decoder layer. Talking specifically about encoder_attention, general logic behind the module was totally fine - if it's our first decoding step and incr_state is still None, values are calculated from encoder_state tensors, otherwise we can just reuse the ones we calculated before. But there was a bug in implementation that forced MultiHeadAttention to recalculate this values despite replacing them with completely similar values saved in incr_state afterwards. So it went like "Let's get processed encoder tensors - Oh, nevermind, we've already had them". It's quite an obvious thing, but it doesn't strike the eye much unless you specifically check model logic to look for the ways to optimize latency