@dejk-alg 2021-03-26
If you are interested, I've spent a bit of time with blender model, and while setting I work with is pretty different from what you work with here in the lib, I think I could add a couple of optimization changes to transformer modules. But all of them would generally be longer in terms of code and not that large in terms of latency decrease. So I'd love to know whether you think it's worth it, given that this changes will have to be reviewed by you or your team anyway